change focus from government to governance
There are three essential sectors of civic participation/forces: the state/government, the market/business, and voluntarism/nonprofit. Many activities and actors comprise Code for America as a member-based organization, and there are many projects that include non-state entities as key actors. If we include in our values the word "government" as it refers to the federal, state, and local, we should qualify that it does not mean our goals are entirely oriented toward a centralized public service vendor. If, instead, we focus on "governance," then that is more inclusive of citizen politics, nonprofit entities, and peer-to-peer organization.