Added operating principle "Fairness for all"
Legislation, executive orders, and decisions made by agencies have far-reaching effects that—under the current administration—will challenge fairness for current and prospective citizens, residents, and visitors to our country. The decisions we make as a community of technology workers in civic tech must reflect our country's founding principles. Decisions that threaten that based on short-sightedness, malice, or ignorance must be checked thoroughly and reckoned with by the community if we truly want to change how our government works.